+++ src/components/Layout/index.ts
@@ -1,2 +1,3 @@
 export { Layout } from './Layout';
+export { LayoutHeader } from './LayoutHeader';
 export { LayoutContext, LayoutProvider } from './LayoutProvider';
